'check' renamed to 'doctor', 'telescope' commands moved to separate file
